Track commanders' ships

+++ src/modules/logs/content/log_event_content/load_game_event.rs
@@ -30,7 +30,7 @@ pub struct LoadGameEventShipInfo {
     pub ship: ShipType,
 
     #[serde(rename = "ShipID")]
-    pub ship_id: u32,
+    pub ship_id: u64,
     pub ship_name: String,
     pub ship_ident: String,
     pub fuel_level: f32,
